Update styles.css
George edited a lot of css to change the colour scheme. He experimented with an image for a background, and then decided it was too much and tried a gradient background which look more appealing. Additionally i began to comment the css to make it easier to talk about a section of text.