Update form.php
over the past few days i have been struggling to make the database connection work with my website. Once i fixed this (the form has to be filled out and then that information gets sent to our database) i decide to echo the information from the database back to the website so it's similar to a live news feed so it becomes interactive and almost live.